refactor, add LICENSE.txt
