+++ .gitignore
@@ -194,6 +194,8 @@ yarn-error.log*
 
 logs
 
+.aider*
+
 # agent
 .envrc
 /workspace
